Initialize project umbrella
Add umbrella for frontend, IoT and core app. Initially the project was
monolithic with support for iot and frontend. However, it needs to
be broken to be inline with single responsibility principle.

The commit adds all the previous test files and moves configuration
for all the apps to umbrella level config as mentioned for
umbrella from elixir lang core creators.
Link:
https://elixir-lang.org/getting-started/mix-otp/dependencies-and-umbrella-projects.html

The previous commits for the featurs added till this commit
can be found at the link below.
https://github.com/DataKrewTech/AcqDat