Add Signout and Refresh Token (#3)
## Why
The access token needs to be renewed as it has a short expiry time, this can be handled with a refresh token with a long expiry time. Upon expiry `refresh token` can be exchanged for a new access token. Upon sign out both access token and refresh token should be revoked.

### Changes
- Add APIs to refresh or renew the access token.
- Add APIs signout