Add hierarchy v5 (#15)
Why?
The old hierarchy, based on sites, processes and sections, wasn't flexible enough to absorb industry-wide use cases. In order to cover various use cases, a new hierarchy was decided.
The new hierarchy removes the concept of sites and processes.

New Design
The new design consists of the following:

Entity: An entity could be an organization, asset, sensor or gateway.
Organization: An organization is defined as a client we have on-boarded
Asset: An asset is defined as any property, object, space of the organization. It can be static or mobile. Examples are factory, truck, equipment, container, treatment plant, solar panel, data center, building, etc.
Sensor: A sensor is a sensor as defined by [Wikipedia] (https://en.wikipedia.org/wiki/Sensor)
Gateway: A gateway is what is used to capture data from a sensor and upload it to a cloud server.