User invitations (#20)
# Why?
In order to allow the user to have access to this app, they need to be invited by admin/others.
Admin will invite the user. Upon invitation, the user will receive the invitation email along with a link. On clicking the link, the user will be redirected to a page where they'll enter password and user information.

## Relations
We maintain an invitations table, which will store all the details of the invited user.
Afterward, when user data is created, the data from the invitations table will be removed.

## Changes
- Create Invitations Table, Migrations and Schema
- Create Invitations Model
- Add Invite API
- Use Sendgrid for Emails
- Create Emails Template
- Adde Create Password/User API
- Add user to elastic search