Domain logic handling for adding asset_type (#43)
# Why?
Asset Types allows grouping of assets. The user should be able to create asset types in the system so it can be used while creating assets. Also, there was a need to ease `sensor_type` creation from the asset_type parameters. The user would now be able to generate sensor_type while adding an `asset_type`. The asset_type metadata is referred to while setting the metadata of an asset referencing it.

## Changes
-  Add field to allow asset type to hold parameters for streaming data
-  Add a field for adding metadata to the asset type.
-  Add check to see if asset type holds parameter for streaming data.
-  Add flow to create a sensor type from asset type if it's marked for storing streaming data.
-  Add API to create an Asset Type (will create a sensor type if flag checked).
-  The value of these metadata fields should be stored while creating an asset.
-  Add a field for handling sensor type user-generated or from asset type generated, using `generated_by`.
- Make use of UUIDs for storing metadata of asset type in assets.
- Make use of UUIDs in sensor type parameters also.
- By default map sensor parameters to asset parameters.