Move path evaluation to runtime in app seed
Earlier the path for `acqdat_core` was held in a module attribute,
this led to compile time evaluation of the path which mismatched
when it was used during runtime. During release creation, app
paths are different from the compile time. This led to an error
in reading the files during seed operation.