Hierarchy and data dump for IoT manager (#55)
# Why?
The user should be able to see the project hierarchy along with gateways. Another requirement is to set the JSON spec which would be used while parsing data received from a gateway. The data dump should be handled in a seamless manner so as not to block the calling API.

## Changes
- Add API to load gateway hierarchy
- Add schema and database for dumping data received from a gateway.
- Improve folder structure for handling IoT manager files.
- Seed file modification to incorporate the new changes in gateway