Data Dump and Parser implementation with workers (#58)
# Why
In order to handle data uploads from gateways and then parse them to create respective data for gateway and sensor worker processes are used. The worker processes parallelize the task of dumping and parsing data leading to less overhead in the API call. Also, there was a need for HTTP API to upload the data.

##  Changes
- Add Data dump table schema and migration.
- Add API to upload data to the server
- Add worker pool to dump data received by the server from a gateway in the database.
- Add Worker pool to handle parsing data from data dump and storing it in sensors data and gateway data records.

Othe Changes
- Create a search for users of an organization. 
- Add support for elastic search in GitHub actions.

Co-authored-by: Ayoush <ayoush@Ayoushs-MacBook-Air.local>
Co-authored-by: arjun289 <arjun289singh@gmail.com>