Add data crunching support (#74)
# Why?
In order to provide a flexible data crunching capability in the system, a data cruncher app has been added. The data cruncher app makes use of flow-based programming to analyze the data received from the sensor. The functions being used for analyzing the data are provided as components that can be connected together to transform data.
The flows are modeled as tasks that can be executed. A task consists of multiple workflows which are essentially directed acyclic graphs in which every node in the graph transforms the data and passes it to the next one. To create a truly concurrent execution environment Virta, an elixir library was used as the base upon which modifications have been made.

## Changes
- Add Virta as an app in umbrella
- Add schema for tasks
- Add functions for time series univariate
- Add migration for tasks table
- Add skeleton for workflow and task modules
- Add token passing in functions
- Add table workflows
  - Each task can have multiple workflows, through has_may associations
  - Each workflow consists of the following
     - graph(denoting the flow)
     - input_data(to initiate the flow with)
- Implement flow for the workflow registrations inside the task
- Implement execute flow for the registered workflows
- Add the temp_workflow output, in the task execute API flow
- API to list all tasks, API to return a task detail
- API to register and execute a task.
- Add WebSockets support to show the data when a task has been completed.

Co-authored-by: arjun289 <arjun289singh@gmail.com>