Memory overflow issue when generating Analysis Table (#172)
#Why?
A certain arrangement of the asset tree caused the data insights sub tree generation code to run into infinite recursion
causing a memory overflow. The reason behind it is using the DB ids which could be the same for a parent and child node. It is being handled now by making the id a combination of name and id so it remains unique.

## Changes
- Add N Ary tree data structure
- Make the N-ary tree node structure a combination of the asset name and id