Auto signup and invitation changes (#188)
# Why?
Users should be able to register themselves, upon registration a new organization is created with the user as org-admin.  When a user registers a request is created, the super-admin can decide to accept or reject the request. Until now all the registration and organization creation was being added by the back office. Also, a user can belong to multiple organizations with different roles. Upon invitation flow should be handled properly as a user who is already in the system can be invited to another organization. 
The roles existing in the system have also been modified in the following way:
- admin -> super-admin
- manager -> org-admin
- user -> user
Changes in elasticsearch to cater to the new user credentials table.

## Changes
- Change the tenant manager logic of adding the user as admin in the organization. Now  the user will be added 
  as org-admin
- Change the user table and moved the user credentials to the new user credentials table, changed the user schema
- Modify sign_in API flow to take org_id along with email and password.
- Add new column type to invitations table, as well as modify the index criteria for email.
- Modify the existing invitation flow, to incorporate respective user invitations to multiple organizations.
- Modify user seed to plan according to the current user organization relation
- Add logic to migrate the data correctly
- Fix rollback data seeding for user and user credentials
- Fix rollback data seeding for user and user credentials
- Modify user creation flow, after they accept the invitation
- Change the user apps and asset association
- Add separate API to do org sign-in
- Modify role names
- Add validate token auth plug to validate auth_token
- Add API to list all the registration requests
- Add API to update request status to approved after sending invitation
- Add migration to associate user_settings with user_credentials
- Add migration to move avatar column from users table to user_credentials table
- Add API to show user credentials and update user credentials

Co-authored-by: Ayoush <ayoushchourasia1418@gmail.com>