Member deletion error in role manager (#177)
## Change
A member (user) is able to delete other members. A member is also able to delete the admin. After deleting a user, it’s still possible for that deleted user to make requests. A user should not be able to delete other members and the admin, also once the user is removed the user requests should not be authorized.