Parameter Mapping enhancements (#201)
# Why?
A sensor can be connected to only one gateway at a time. When a sensor's parameters are mapped to the gateway key,  the sensor is connected to the gateway. If parameters of a sensor connected to the existing gateway are mapped to a new gateway then the connection is severed with the old gateway and a new connection is formed. All the mappings from the older gateways are removed when this process happens.