Add support for grouping in alerts (#239)
## Why?
Alerts are used for notifying a user if certain conditions are met in the platform. The conditions are defined using alert rules.
At present every time the alert rule is met, a notification is met. This is undesirable as too many emails would be sent if a sensor is meeting some threshold rule. In order to take care of this, alerts are grouped together by time. The PR introduces a way to allow groupings of alerts.
In order to group the alerts together, a grouping technique is used based on MurmurHash2. Every alert is grouped together based on certain properties:
- Alert policy
- App Name
- Entity Id
- Entity Name

## Changes
- Modify the alert DB and schema to introduce alerts and alert event log.
- Add support for grouping with the help of murmur hash.
- Add rules based on time-based grouping.
- Modify the alert creation logic to create a token and pass it to the grouping module.
- Modify the concurrent task creation logic for handling alerts with Task.Supervisor.

Co-authored-by: arjun289 <arjun289singh@gmail.com>