Refactor MQTT subscription and data parsing logic (#252)
# Why?
The MQTT subscription per project has been removed as it was causing issues. The new subscription is server-based and there is only one MQTT TCP connection to handle incoming data, this will be improved in the future by creating shared subscriptions to do load balancing properly.
The old Genserver based data dump and parsing have been removed as it was not scalable and had certain issues. The new data dump logic is a simple design based on the GenStage producer and consumer supervisor approach, where the internal GenStage buffer is being used for queueing incoming events and the consumer supervisor is being used to start workers to handle the incoming events.

## Changes
- Add support for MQTT to test Github actions.
- Remove per project subscription
- Add support for server-level subscription
- Remove GenServer based IoT data dump and parsing logic
- Add GenStage based support for dump and parsing logic